Solved Validate BST using Recursive Inorder Traversal
O(n) time and O(n) space (for the call-stack). Instead of storing the inorder traversal in a list and seeing if that list was strictly increasing, I simply kept comparing whether the current element in the inorder traversal was greater than the previous element in the inorder traversal; this reduces this portion of the computation from O(n) memory to O(1) memory.